Temporary save experimental changes

+++ hammer/mapdecal.cpp
@@ -432,7 +432,7 @@ void CMapDecal::OnNotifyDependent(CMapClass *pObject, Notify_Dependent_t eNotify
 		// We could not re-attach to the solid because it was moved out of range or deleted. If we are
 		// no longer attached to any solids, remove our entity from the world.
 		//
-		int index = m_Solids.Find(pSolid);
+		intp index = m_Solids.Find(pSolid);
 		if (index != -1)
 		{
 			m_Solids.Remove(index);
