+++ vgui2/vgui_controls/TreeViewListControl.cpp
@@ -86,7 +86,7 @@ void CTreeViewListControl::SetNumColumns( int nColumns )
 	InvalidateLayout();
 }
 
-int CTreeViewListControl::GetNumColumns() const
+intp CTreeViewListControl::GetNumColumns() const
 {
 	return m_Columns.Count();
 }
